pyproject.toml
[tool.poetry]
name = "analytics-app"
version = "0.1.0"
description = ""
authors = ["SahilMehdiyev <branchsahil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-jazzmin = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
